backend/requirements.txt
# Production Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Image Processing & Computer Vision
Pillow==10.3.0
opencv-python<=4.6.0.66

# Deep Learning & ML
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
timm==0.9.12

# Vector Search & Similarity
faiss-cpu==1.7.4
numpy==1.24.4

# Data Handling
pandas==2.0.3

# Progress & Utilities
tqdm==4.66.1

# HTTP & CORS
python-multipart==0.0.6

# Performance optimizations
accelerate==0.24.1
tokenizers==0.15.0
imageio==2.31.5

# HEIF/AVIF support
pillow-heif
pillow-avif-plugin

# OCR (Optional - remove if not used to reduce image size)
paddlepaddle>=2.4.0
paddleocr==2.7.3

# Optional: Redis client for caching
redis==5.0.1

# Optional: Prometheus monitoring
prometheus-client==0.19.0

# Security
cryptography>=41.0.0